fix:invalid type uint256 in fee

+++ src/MangroveGhostBook.sol
@@ -79,7 +79,7 @@ contract MangroveGhostBook is ReentrancyGuard, Ownable {
     if (bestOfferId == 0) {
       externalMaxTick = maxTick;
     } else {
-      uint256 fee = MGV.local(olKey).fee();
+      int256 fee = int256(MGV.local(olKey).fee());
       Tick offerTick = Tick.wrap(Tick.unwrap(MGV.offers(olKey, bestOfferId).tick()) + fee);
       externalMaxTick = Tick.unwrap(offerTick) < Tick.unwrap(maxTick) ? offerTick : maxTick;
     }
